cleaned up vector classes

+++ GLOBALVARIABLES.h
@@ -1,8 +1,8 @@
 #pragma once
 
-float _WINDOW_H;
-float _WINDOW_W;
+int _WINDOW_H;
+int _WINDOW_W;
 
 float _DISPLAY_SCALE;
-double _KEYSIZE;
+float _KEYSIZE;
 float _TEXT_SCALE;
